jamatapp/requirements.txt
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.3
django-toolbelt==0.0.1
djangorestframework==3.5.3
gunicorn==19.6.0
psycopg2==2.6.2
static3==0.7.0
whitenoise==3.2